requirements.txt
-e git://github.com/aio-libs/aiosmtpd#egg=aiosmtpd
atpublic==1.0
pyslack==0.4.0
PyYAML==3.12
requests==2.11.1
pytest==4.4.1
html2text
